Backend/requirements.txt
asgiref
attrs 
cachetools 
certifi 
charset-normalizer 
Django 
django-adminlte-3 
django-adminlte2 
django-cors-headers 
django-dotenv 
django-filter 
django-jazzmin 
django-material-admin 
django-multiselectfield 
djangorestframework 
djangorestframework-simplejwt 
google-auth
idna 
importlib-metadata
importlib-resources
jsonschema 
Markdown
mpmath
#mysqlclient 
packaging
pillow
pkgutil_resolve_name
pyasn1
pyasn1_modules
pycryptodome
PyJWT
pyproject-toml
pyrsistent
pytesseract
pytz
requests
rsa
setuptools
sqlparse 
sympy
toml 
typing_extensions 
tzdata 
urllib3 
wheel 
zipp